requirements.txt
numpy>=1.26
pandas>=2.2
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-genai>=0.3.0
streamlit
matplotlib
tabulate